- updated conditions for deleting an entry - entries were originally deleted if there was no mood and no text, now they are deleted only if there is no mood, no text, and no mentions

+++ app/src/main/java/com/example/gratitudejournal/Alarm.java
@@ -140,11 +140,11 @@ public class Alarm extends BroadcastReceiver {
                     e.printStackTrace();
                 }
 
-                if(mood.equals("skip") && text.equals("No entry")){
+                if(mood.equals("skip") && text.equals("No entry") && fm.length() == 0 && cfm.length() == 0){
                     lastEntry.deleteInBackground();
                 }
 
-                if(mood.equals("No mood selected") && text.equals("No entry")){
+                if(mood.equals("No mood selected") && text.equals("No entry") && fm.length() == 0 && cfm.length() == 0){
                     lastEntry.deleteInBackground();
                 }
